crypto-auditing: Define ContextTracker at the top-level crate
