+++ crypto-auditing/src/lib.rs
@@ -1,5 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (C) 2022-2023 The crypto-auditing developers.
 
+mod context_tracker;
 pub mod event_broker;
 pub mod types;
+pub use context_tracker::ContextTracker;
